requirements/requirements-dev.txt
# SHA1:6f2f173537fc654cc86d86586d9ec283873eb38d
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r requirements-base.txt
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via
    #   aiohttp-jinja2
    #   textual-dev
    #   textual-serve
aiohttp-jinja2==1.6
    # via textual-serve
aiosignal==1.3.2
    # via aiohttp
attrs==25.1.0
    # via aiohttp
cfgv==3.4.0
    # via pre-commit
coverage==7.6.10
    # via -r requirements/requirements-dev.in
distlib==0.3.9
    # via virtualenv
filelock==3.17.0
    # via virtualenv
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
identify==2.6.8
    # via pre-commit
idna==3.10
    # via yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via
    #   aiohttp-jinja2
    #   pytest-textual-snapshot
    #   textual-serve
markupsafe==3.0.2
    # via jinja2
msgpack==1.1.0
    # via textual-dev
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nodeenv==1.9.1
    # via
    #   pre-commit
    #   pyright
pluggy==1.5.0
    # via pytest
pre-commit==4.1.0
    # via -r requirements/requirements-dev.in
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pyright==1.1.392.post0
    # via -r requirements/requirements-dev.in
pytest==8.3.4
    # via
    #   -r requirements/requirements-dev.in
    #   pytest-asyncio
    #   pytest-textual-snapshot
    #   syrupy
pytest-asyncio==0.25.2
    # via -r requirements/requirements-dev.in
pytest-textual-snapshot==1.0.0
    # via -r requirements/requirements-dev.in
pyyaml==6.0.2
    # via pre-commit
ruff==0.9.3
    # via -r requirements/requirements-dev.in
syrupy==4.8.2
    # via pytest-textual-snapshot
textual-dev==1.7.0
    # via -r requirements/requirements-dev.in
textual-serve==1.1.1
    # via textual-dev
virtualenv==20.29.2
    # via pre-commit
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
